Fix feature combinations
