+++ porky/src/porky.rs
@@ -7,7 +7,6 @@ use crate::pin_descriptions::PIN_DESCRIPTIONS;
 use crate::ssid::{MARKER_LENGTH, SSID_NAME, SSID_NAME_LENGTH, SSID_PASS, SSID_PASS_LENGTH};
 use core::str;
 use cyw43_pio::PioSpi;
-//noinspection RsUnusedImport
 use defmt::{error, info};
 use defmt_rtt as _;
 use embassy_executor::Spawner;
@@ -16,12 +15,14 @@ use embassy_rp::bind_interrupts;
 use embassy_rp::flash::Async;
 use embassy_rp::flash::Flash;
 use embassy_rp::gpio::{Level, Output};
+#[cfg(feature = "usb")]
 use embassy_rp::peripherals::USB;
 use embassy_rp::peripherals::{DMA_CH1, FLASH, PIO0};
 use embassy_rp::pio::InterruptHandler as PioInterruptHandler;
 use embassy_rp::pio::Pio;
-#[cfg(feature = "usb-tcp")]
+#[cfg(feature = "usb")]
 use embassy_rp::usb::Driver;
+#[cfg(feature = "usb")]
 use embassy_rp::usb::InterruptHandler as USBInterruptHandler;
 use embassy_sync::blocking_mutex::raw::ThreadModeRawMutex;
 use embassy_sync::channel::Channel;
@@ -31,6 +32,11 @@ use heapless::Vec;
 use panic_probe as _;
 use static_cell::StaticCell;
 
+#[cfg(all(feature = "usb-tcp", feature = "usb-raw"))]
+compile_error!(
+    "Features 'usb-raw' and 'usb-tcp' are mutually exclusive and cannot be enabled together"
+);
+
 /// The ssid config generated by build.rs in "$OUT_DIR/ssid.rs"
 mod ssid {
     include!(concat!(env!("OUT_DIR"), "/ssid.rs"));
@@ -38,6 +44,9 @@ mod ssid {
 
 /// Wi-Fi related functions
 mod wifi;
+
+#[cfg(feature = "usb")]
+mod usb;
 
 #[cfg(feature = "usb-tcp")]
 /// Module for Tcp over USB
@@ -63,6 +72,7 @@ const FLASH_SIZE: usize = 2 * 1024 * 1024;
 
 bind_interrupts!(struct Irqs {
     PIO0_IRQ_0 => PioInterruptHandler<PIO0>;
+    #[cfg(feature = "usb")]
     USBCTRL_IRQ => USBInterruptHandler<USB>;
 });
 
@@ -168,16 +178,18 @@ async fn main(spawner: Spawner) {
     let serial_number = serial_number(peripherals.FLASH, peripherals.DMA_CH1);
     let hw_desc = hardware_description(serial_number);
 
-    //    start_usb(spawner, driver, "12345678").await;
+    #[cfg(feature = "usb")]
+    let driver = Driver::new(peripherals.USB, Irqs);
 
     #[cfg(feature = "usb-tcp")]
-    let driver = Driver::new(peripherals.USB, Irqs);
-    #[cfg(feature = "usb-tcp")]
-    let usb_stack = usb_tcp::start_net(spawner, driver, serial_number).await;
+    let usb_stack = usb_tcp::start(spawner, driver, serial_number).await;
     #[cfg(feature = "usb-tcp")]
     let mut usb_tx_buffer = [0; 4096];
     #[cfg(feature = "usb-tcp")]
     let mut usb_rx_buffer = [0; 4096];
+
+    #[cfg(feature = "usb-raw")]
+    usb_raw::start(spawner, driver, serial_number).await;
 
     let ssid_name = SSID_NAME[MARKER_LENGTH..(MARKER_LENGTH + SSID_NAME_LENGTH)].trim();
     let ssid_pass = SSID_PASS[MARKER_LENGTH..(MARKER_LENGTH + SSID_PASS_LENGTH)].trim();
