+++ porky/src/usb_tcp.rs
@@ -1,3 +1,4 @@
+use crate::usb::get_usb_builder;
 use defmt::*;
 use embassy_executor::Spawner;
 use embassy_net::{Stack, StackResources};
@@ -6,7 +7,7 @@ use embassy_rp::peripherals::USB;
 use embassy_rp::usb::Driver;
 use embassy_usb::class::cdc_ncm::embassy_net::{Device, Runner, State as NetState};
 use embassy_usb::class::cdc_ncm::{CdcNcmClass, State};
-use embassy_usb::{Builder, Config, UsbDevice};
+use embassy_usb::UsbDevice;
 use rand::RngCore;
 use static_cell::StaticCell;
 use {defmt_rtt as _, panic_probe as _};
@@ -30,40 +31,8 @@ async fn net_task(mut runner: embassy_net::Runner<'static, Device<'static, MTU>>
     runner.run().await
 }
 
-pub(crate) fn get_usb_builder(
-    driver: Driver<'static, USB>,
-    serial: &'static str,
-) -> Builder<'static, Driver<'static, USB>> {
-    // Create embassy-usb Config
-    let mut config = Config::new(0xbabe, 0xface);
-    config.manufacturer = Some("pigg");
-    config.product = Some("porky");
-    config.serial_number = Some(serial);
-    config.max_power = 100;
-    config.max_packet_size_0 = 64;
-
-    // Required for Windows support.
-    config.device_class = 0xEF;
-    config.device_sub_class = 0x02;
-    config.device_protocol = 0x01;
-    config.composite_with_iads = true;
-
-    static CONFIG_DESC: StaticCell<[u8; 256]> = StaticCell::new();
-    static BOS_DESC: StaticCell<[u8; 256]> = StaticCell::new();
-    static CONTROL_BUF: StaticCell<[u8; 128]> = StaticCell::new();
-    static MSOS_DESC: StaticCell<[u8; 256]> = StaticCell::new();
-    Builder::new(
-        driver,
-        config,
-        &mut CONFIG_DESC.init([0; 256])[..],
-        &mut BOS_DESC.init([0; 256])[..],
-        &mut MSOS_DESC.init([0; 256])[..],
-        &mut CONTROL_BUF.init([0; 128])[..],
-    )
-}
-
 /// Start a network stack based on USB
-pub async fn start_net(
+pub async fn start(
     spawner: Spawner,
     driver: Driver<'static, USB>,
     serial: &'static str,
